Revert "DB-8570: invalidate dictionary cache after updating system procedures(2.8) (#2841)"
This reverts commit c26cf99fb8833cd84ce01cf8b785bd7c34b883e8.
