+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/catalog/DataDictionaryCache.java
@@ -316,9 +316,6 @@ public class DataDictionaryCache {
         spsNameCache.put(tableKey, sps);
     }
 
-    public void clearSpsNameCache() throws StandardException {
-        spsNameCache.invalidateAll();
-    }
     public void storedPreparedStatementCacheAdd(SPSDescriptor desc) throws StandardException {
         if (!dd.canWriteCache(null))
             return;
@@ -335,9 +332,6 @@ public class DataDictionaryCache {
         return storedPreparedStatementCache.getIfPresent(uuid);
     }
 
-    public void clearStoredPreparedStatementCache() {
-        storedPreparedStatementCache.invalidateAll();
-    }
 
     public Conglomerate conglomerateCacheFind(TransactionController xactMgr,Long conglomId) throws StandardException {
         if (!dd.canReadCache(xactMgr) && conglomId>=DataDictionary.FIRST_USER_TABLE_NUMBER)
@@ -478,10 +472,6 @@ public class DataDictionaryCache {
         defaultRoleCache.invalidateAll();
         roleGrantCache.invalidateAll();
         tokenCache.invalidateAll();
-    }
-
-    public void clearAliasCache() {
-        aliasDescriptorCache.invalidateAll();
     }
 
     public void clearSchemaCache(){
