+++ splice_machine/src/main/java/com/splicemachine/derby/impl/sql/catalog/SpliceSystemProcedures.java
@@ -1392,22 +1392,6 @@ public class SpliceSystemProcedures extends DefaultSystemProcedureGenerator {
                             .build();
                     procedures.add(disableTableReplication);
 
-                    Procedure updateAllSystemProcedures = Procedure.newBuilder()
-                            .name("SYSCS_UPDATE_ALL_SYSTEM_PROCEDURES")
-                            .numOutputParams(0).numResultSets(0).modifiesSql()
-                            .returnType(null).isDeterministic(false)
-                            .ownerClass(SpliceAdmin.class.getCanonicalName())
-                            .build();
-                    procedures.add(updateAllSystemProcedures);
-
-                    Procedure updateSystemProcedure = Procedure.newBuilder().name("SYSCS_UPDATE_SYSTEM_PROCEDURE")
-                            .numOutputParams(0).numResultSets(0).modifiesSql()
-                            .returnType(null).isDeterministic(false)
-                            .ownerClass(SpliceAdmin.class.getCanonicalName())
-                            .catalog("schemaName")
-                            .catalog("procName")
-                            .build();
-                    procedures.add(updateSystemProcedure);
                 }  // End key == sysUUID
 
             } // End iteration through map keys (schema UUIDs)
