+++ splice_machine/src/main/java/com/splicemachine/derby/impl/sql/catalog/upgrade/UpgradeScriptBase.java
@@ -26,7 +26,7 @@ import com.splicemachine.derby.impl.sql.catalog.SpliceDataDictionary;
 /**
  * Created by jyuan on 10/17/14.
  */
-public class UpgradeScriptBase implements UpgradeScript {
+public abstract class UpgradeScriptBase implements UpgradeScript {
 	protected static final Logger LOG = Logger.getLogger(UpgradeScriptBase.class);
 
     protected SpliceDataDictionary sdd;
