+++ splice_machine/src/main/java/com/splicemachine/derby/utils/SpliceAdmin.java
@@ -61,7 +61,6 @@ import com.splicemachine.pipeline.Exceptions;
 import com.splicemachine.pipeline.SimpleActivation;
 import com.splicemachine.protobuf.ProtoUtil;
 import com.splicemachine.si.api.data.TxnOperationFactory;
-import com.splicemachine.si.api.txn.TxnView;
 import com.splicemachine.si.impl.driver.SIDriver;
 import com.splicemachine.storage.*;
 import com.splicemachine.utils.Pair;
@@ -2450,66 +2449,4 @@ public class SpliceAdmin extends BaseAdminProcedures{
         }
         return false;
     }
-
-    /**
-     * Create or update all system stored procedures.  If the system stored procedure alreadys exists in the data dictionary,
-     * the stored procedure will be dropped and then created again.
-     *
-     * @throws SQLException
-     */
-    public static void SYSCS_UPDATE_ALL_SYSTEM_PROCEDURES()
-            throws SQLException{
-        try{
-            LanguageConnectionContext lcc=ConnectionUtil.getCurrentLCC();
-            TransactionController tc=lcc.getTransactionExecute();
-            DataDictionary dd=lcc.getDataDictionary();
-
-            dd.startWriting(lcc);
-            TxnView activeTransaction = ((SpliceTransactionManager) tc).getActiveStateTxn();
-            DDLMessage.DDLChange ddlChange = ProtoUtil.createUpdateSystemProcedure(activeTransaction.getTxnId());
-            tc.prepareDataDictionaryChange(DDLUtils.notifyMetadataChange(ddlChange));
-
-            dd.createOrUpdateAllSystemProcedures(tc);
-
-        }catch(StandardException se){
-            throw PublicAPI.wrapStandardException(se);
-        }
-    }
-
-    /**
-     * Create or update a system stored procedure.  If the system stored procedure alreadys exists in the data dictionary,
-     * the stored procedure will be dropped and then created again.
-     *
-     * @param schemaName name of the system schema
-     * @param procName   name of the system stored procedure
-     * @throws SQLException
-     */
-    public static void SYSCS_UPDATE_SYSTEM_PROCEDURE(String schemaName,String procName)
-            throws SQLException{
-        try{
-            LanguageConnectionContext lcc=ConnectionUtil.getCurrentLCC();
-            TransactionController tc=lcc.getTransactionExecute();
-            DataDictionary dd=lcc.getDataDictionary();
-
-            /*
-            ** Inform the data dictionary that we are about to write to it.
-            ** There are several calls to data dictionary "get" methods here
-            ** that might be done in "read" mode in the data dictionary, but
-            ** it seemed safer to do this whole operation in "write" mode.
-            **
-            ** We tell the data dictionary we're done writing at the end of
-            ** the transaction.
-            */
-            dd.startWriting(lcc);
-
-            TxnView activeTransaction = ((SpliceTransactionManager) tc).getActiveStateTxn();
-            DDLMessage.DDLChange ddlChange = ProtoUtil.createUpdateSystemProcedure(activeTransaction.getTxnId());
-            tc.prepareDataDictionaryChange(DDLUtils.notifyMetadataChange(ddlChange));
-
-            dd.createOrUpdateSystemProcedure(schemaName,procName,tc);
-        }catch(StandardException se){
-            throw PublicAPI.wrapStandardException(se);
-        }
-    }
-
 }
