+++ splice_protocol/src/main/protobuf/DDL.proto
@@ -307,8 +307,6 @@ message NotifyModifyClasspath {
     optional string classpath = 1;
 }
 
-message UpdateSystemProcedures{
-}
 
 enum DDLChangeType {
     CHANGE_PK = 0;
@@ -352,7 +350,6 @@ enum DDLChangeType {
     UPDATE_SCHEMA_OWNER=38;
     GRANT_REVOKE_ROLE = 39;
     SET_DATABASE_PROPERTY = 40;
-    UPDATE_SYSTEM_PROCEDURES = 41;
 }
 
 message DDLChange {
@@ -393,5 +390,4 @@ message DDLChange {
     optional UpdateSchemaOwner updateSchemaOwner = 35;
     optional GrantRevokeRole grantRevokeRole = 36;
     optional SetDatabaseProperty setDatabaseProperty = 37;
-    optional UpdateSystemProcedures updateSystemProcedures = 38;
 }
